replace old docs pages with redirects to new ones
Closes everypolitician/everypolitician#437

We've moved the documentation pages over to http://docs.everypolitician.org
but the old ones remain in this repo; there might be some incoming links
to them, so rather than just deleting, I've replaced them with placeholder
pages with `meta refresh` tags in their headers.

This isn't a great solution but it's the best we can currently do on
GitHub Pages (since there's no .htaccess or request-level redirect
available). Note that I haven't used GitHub's `redirect-from` capability
because all that's doing is injecting such `meta` HTTP-refresh tags anyway.
Their solution is handy if you're generating pages from Jekyll but as
we're not it's just as easy to drop these placeholders in. Maybe I
could have seperated these pages out as a template but that seems
overkill since they're deprecated anyway..

Note that all the pages redirect to their identically-named file
on docs.everypolitician.org _except_ `about.html` which redirects to
`index.html` on that subdomain.

Also, in case it's not clear, some of these old pages' text was out
of date (of course) which is why I've replaced their contents with
placeholders (lifted from 404.html).